refactor!: expose HelpInfo as a namedtuple (#37)
We refactor the HelpInfo type to be a namedtuple and be part of
the public API for better consumption of the information. We also
drop the assumption that the variable name would be formatted in
a RST document and remove the double backtick wrapping from the
generated help info data.